modularize html and api routes from server. begin readme file
