+++ README.md
@@ -1,2 +1,8 @@
 # express-noteBook
-UCB Bootcamp Module 11 Homework Challenge
+## UCB Bootcamp Module 11 Homework Challenge
+
+## Review
+Heroku deployment:
+GitHub repository: https://github.com/jmariemo/express-noteBook
+
+##
