+++ routes/htmlRoutes.js
@@ -1,0 +1,12 @@
+const app = require("express");
+const path = require("path");
+
+module.exports = (app) => {
+    app.get("/notes", (req, res) =>
+  res.sendFile(path.join(__dirname, "../public/notes.html"))
+);
+
+app.get("/", (req, res) =>
+  res.sendFile(path.join(__dirname, "../public/index.html"))
+);
+}
